Crate usbd_hid_descriptors

Structs

ItemPrefix

ItemPrefix describes the 1 byte prefix describing an item in a descriptor.

MainItemSetting

MainItemSetting describes the bits which configure invariants on a MainItem.

Enums

GlobalItemKind

GlobalItemKind describes global item tags as described in section 6.2.2.7 ‘Report Descriptor’ of the spec, version 1.11.

ItemType

ItemType describes types of items as described in section 6.2.2.7 ‘Report Descriptor’ of the spec, version 1.11.

LocalItemKind

LocalItemKind describes local item tags as described in section 6.2.2.8 ‘Local Items’ of the spec, version 1.11.

MainItemKind

MainItemKind describes main item tags as described in section 6.2.2.4 ‘Report Descriptor’ of the spec, version 1.11.